Add track count and filtered track count.
